+++ resources/src/router.js
@@ -1,8 +1,9 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 
+// auth
 // import Signup from '@/components/pages/Signup'
-import Login from '@/components/organisms/Login'
+import Login from '@/components/organisms/auth/Login'
 // import Home from '@/components/pages/Home'
 
 Vue.use(Router)
